requirements.txt
agno
python-dotenv
faiss-cpu
pandas
numpy
groq
openai
duckduckgo-search
pygithub